+++ EssentialFeed/EssentialFeediOS/Feed UI/Composers/FeedUIComposer.swift
@@ -36,26 +36,3 @@ private extension FeedViewController {
 		return feedController
 	}
 }
-
-private final class FeedViewAdapter: FeedView {
-	private weak var controller: FeedViewController?
-	private let imageLoader: FeedImageDataLoader
-	
-	init(controller: FeedViewController, imageLoader: FeedImageDataLoader) {
-		self.controller = controller
-		self.imageLoader = imageLoader
-	}
-	
-	func display(_ viewModel: FeedViewModel) {
-		controller?.tableModel = viewModel.feed.map { model in
-			let adapter = FeedImageDataLoaderPresentationAdapter<WeakRefVirtualProxy<FeedImageCellController>, UIImage>(model: model, imageLoader: imageLoader)
-			let view = FeedImageCellController(delegate: adapter)
-			
-			adapter.presenter = FeedImagePresenter(
-				view: WeakRefVirtualProxy(view),
-				imageTransformer: UIImage.init)
-			
-			return view
-		}
-	}
-}
