+++ EssentialFeed/EssentialFeediOS/Feed UI/Composers/FeedViewAdapter.swift
@@ -1,0 +1,29 @@
+//
+//  Copyright © 2019 Essential Developer. All rights reserved.
+//
+
+import UIKit
+import EssentialFeed
+
+final class FeedViewAdapter: FeedView {
+	private weak var controller: FeedViewController?
+	private let imageLoader: FeedImageDataLoader
+	
+	init(controller: FeedViewController, imageLoader: FeedImageDataLoader) {
+		self.controller = controller
+		self.imageLoader = imageLoader
+	}
+	
+	func display(_ viewModel: FeedViewModel) {
+		controller?.tableModel = viewModel.feed.map { model in
+			let adapter = FeedImageDataLoaderPresentationAdapter<WeakRefVirtualProxy<FeedImageCellController>, UIImage>(model: model, imageLoader: imageLoader)
+			let view = FeedImageCellController(delegate: adapter)
+			
+			adapter.presenter = FeedImagePresenter(
+				view: WeakRefVirtualProxy(view),
+				imageTransformer: UIImage.init)
+			
+			return view
+		}
+	}
+}
